pyproject.toml
[tool.poetry]
name = "fastapi-auth-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
psycopg = "^3.1.18"
python-json-logger = "^2.0.7"
sqlmodel = "^0.0.18"
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
mypy = "^1.10.0"

[tool.poetry.group.production.dependencies]
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
